FWSECURITY-1164: Moved retarget_io_pdl into the Platforms directory, updated...

See merge request repo/cy_mcuboot_project/cy_mcuboot!140

+++ boot/cypress/SecureBlinkyApp/main.c
@@ -202,7 +202,7 @@ int main(void)
     app_addr = cy_bl_bnu_policy.bnu_img_policy[1].boot_area.start + CM4_APP_HEADER_SIZE;
 
 #if defined(DEBUG)
-    printf("%s CM4 app address 0x%08x\n\r", GREETING_MESSAGE, app_addr);
+    printf("%s CM4 app address 0x%08x\n\r", GREETING_MESSAGE, (int)app_addr);
     printf("%s Memory regions to protect:\n\r", GREETING_MESSAGE);
 
     for(i = 0; i < POLICY_MAX_N_OF_PROT_REGIONS; i++)
@@ -210,8 +210,8 @@ int main(void)
         if(0 != cy_bl_bnu_policy.prot_regions[i].start)
         {
             printf("address: 0x%08x; size 0x%08x;\n\r",
-                   cy_bl_bnu_policy.prot_regions[i].start,
-                   (uint32_t)(1 << (cy_bl_bnu_policy.prot_regions[i].size + 1)));
+                   (int)cy_bl_bnu_policy.prot_regions[i].start,
+                   (int)(1 << (cy_bl_bnu_policy.prot_regions[i].size + 1)));
         }
     }
 #endif
